abstracted out dnf and apt-get

+++ update-o-matic.sh
@@ -1,23 +1,56 @@
 #!/bin/bash
 
+# BEGIN GlobalVariables
+
+REBOOT=false
+
+# END GlobalVariables
+
 # BEGIN Functions 
 # before the main program, because bash
 
-function linuxFedora {
-	echo Fedora Linux detected
+function executableFileExists {
+	if [ ! -f `which $1` ]
+	then
+		echo $1 does not exist. Please install it before continuing.
+		echo Your package manager may have a way to search for this file
+		exit
+	fi
+}
+
+function rebootIfAllowed {
+	if $REBOOT
+	then
+		if [ -f `which shutdown` ]
+		then
+			shutdown -r +5 "Applying Updates. Cancel with shutdown -c"
+		else
+			echo No shutdown command found.
+		fi
+	else
+		echo Be sure that you do not need to reboot.
+	fi
+}
+
+
+function linuxDnf {
+	executableFileExists dnf
+
+	echo dnf detected
 	dnf check --assumeyes &&
 	dnf autoremove --assumeyes &&
 	dnf upgrade --assumeyes &&
-	shutdown -r +5 "Applying Updates. Cancel with shutdown -c"
+	rebootIfAllowed
 }
 
-function linuxUbuntu {
-	echo Ubuntu Linux detected
+function linuxAptGet {
+	executableFileExists apt-get
+
+	echo Updating using apt-get
 	/usr/bin/apt-get autoremove --yes &&
 	/usr/bin/apt-get update --yes &&
 	/usr/bin/apt-get dist-upgrade --yes &&
-	/sbin/shutdown -r +5 "Applying patches. Cancel with shutdown -c"
-
+	rebootIfAllowed
 }
 
 function darwinMacports {
@@ -57,6 +90,33 @@ function darwinOS {
 
 	echo Please read the messages above and reboot if needed.
 }
+
+function linuxFedora {
+	echo Fedora detected.
+	
+	if [ -f `which dnf` ]
+	then
+		linuxDnf
+	elif [ -f `which yum` ]
+	then
+		echo You are using Yum, which is not yet supported.
+	else
+		Weird. Neither dnf nor yum found. 
+	fi
+}
+
+function linuxUbuntu {
+	echo Ubuntu detected.
+
+	if [ -f `which dnf` ]
+	then
+		linuxAptGet
+	else
+		Weird. apt-get is not installed. Cannot continue.
+		exit.
+	fi
+}
+
 # END Functions
 
 # BEGIN Main
@@ -67,7 +127,7 @@ if [[ $EUID -ne 0 ]]; then
 	exit 1
 fi
 
-
+executableFileExists uname
 majorArch=`uname -s`;
 
 case $majorArch in
@@ -84,10 +144,14 @@ esac
 
 if [ $majorArch = "Linux" ];
 then
+
+	echo Linux detected.
+
+	executableFileExists lsb_release
 	distro=`lsb_release -si`
+
 	case $distro in
-	Fedora)
-		linuxFedora
+	Fedora) linuxFedora
 		;;
 	Ubuntu)
 		linuxUbuntu
